Linux build

+++ src/35_pt/StopPointAddAction.cpp
@@ -104,7 +104,7 @@ namespace synthese
 			// SRID
 			if(_coordinatesSystem)
 			{
-				map.insert(PARAMETER_SRID, _coordinatesSystem->getSRID());
+				map.insert(PARAMETER_SRID, static_cast<int>(_coordinatesSystem->getSRID()));
 			}
 
 			return map;
